* MonoTorrent.Client/Managers/ConnectionManager.cs,   MonoTorrent.Client/ClientEngine.cs: Backport fix for casting a   boxed long
svn path=/branches/bitsharp-0.50/; revision=114327

+++ src/MonoTorrent/MonoTorrent.Client/ClientEngine.cs
@@ -280,7 +280,7 @@ namespace MonoTorrent.Client
         {
             get
             {
-                return (int)Toolbox.Accumulate<TorrentManager>(torrents, delegate(TorrentManager m) { return m.Monitor.DownloadSpeed; });
+                return (int)(long)Toolbox.Accumulate<TorrentManager>(torrents, delegate(TorrentManager m) { return m.Monitor.DownloadSpeed; });
             }
         }
 
@@ -288,7 +288,7 @@ namespace MonoTorrent.Client
         {
             get
             {
-                return (int)Toolbox.Accumulate<TorrentManager>(torrents, delegate(TorrentManager m) { return m.Monitor.UploadSpeed; });
+                return (int)(long)Toolbox.Accumulate<TorrentManager>(torrents, delegate(TorrentManager m) { return m.Monitor.UploadSpeed; });
             }
         }
